🐞 Allow empty job property hashes
The windows-syslog-release added a new job that doesn't have any
properties, but declares an empty `properties:` field in the YAML. This
is valid YAML that was not being parsed by the deprecated candiedyaml
package.

We've switched to a maintained YAML parsing library instead. This
required some attributes to be marked with a name `-` to prevent the
parser from automatically reading YAML based on field name.

Signed-off-by: Brian Upton <bupton@vmware.com>