Remove remnants of deprecated candiedyaml
We've switched the remaining usages of the deprecated candiedyaml
library in favor of the standard yaml library. This revealed a behavior
change where candiedyaml was automatically decoding base64 content,
whereas yaml does not.

We were only ever using some of the "Raw" fields to do this conversion.
So we've removed those Raw fields and just did the deserialization
ourselves.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>